+++ app/Http/Controllers/MahasiswaController.php
@@ -4,12 +4,18 @@ namespace App\Http\Controllers;
 
 use App\Imports\MahasiswaImport;
 use App\Models\Alumni;
+use App\Models\BeritaKomentar;
+use App\Models\Diskusi;
+use App\Models\DiskusiTanyaJawab;
+use App\Models\LokerTanyaJawab;
 use App\Models\Mahasiswa;
 use App\Models\User;
 use Illuminate\Validation\Rules;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Storage;
 use Maatwebsite\Excel\Facades\Excel;
+use Intervention\Image\ImageManagerStatic as Image;
 
 class MahasiswaController extends Controller
 {
@@ -46,16 +52,16 @@ class MahasiswaController extends Controller
     public function store(Request $request)
     {
         $rules = [
-            'npm' => 'required|string|max:9',
+            'npm' => 'required|string|max:9|min:9',
             'nama' => 'required|string|max:40',
             'agama' => 'required|string|max:10',
             'tempat_lahir' => 'required|string|max:20',
             'tanggal_lahir' => 'required|date',
             'jenis_kelamin' => 'required|in:L,P',
-            'alamat' => 'required|string|max:50',
+            'alamat' => 'required|string|max:255',
             'nama_ayah' => 'required|string|max:40',
             'nama_ibu' => 'required|string|max:40',
-            'alamat_orang_tua' => 'required|string|max:50',
+            'alamat_orang_tua' => 'required|string|max:255',
             'pekerjaan_ayah' => 'required|string|max:20',
             'pekerjaan_ibu' => 'required|string|max:20',
             'no_hp' => 'required|numeric',
@@ -71,8 +77,9 @@ class MahasiswaController extends Controller
             'required' => 'Field :attribute wajib diisi',
             'string' => 'Field :attribute harus berupa string',
             'numeric' => 'Field :attribute harus berupa angka',
-            'max' => 'Field :attribute maksimal :size',
-            'digits' => 'Field :attribute maksimal :value digit',
+            'max' => 'Field :attribute maksimal :max',
+            'min' => 'Field :attribute maksimal :min',
+            'digits' => 'Field :attribute harus :digits digit',
             'email' => 'Field :attribute harus berupa email',
             'unique' => 'Field :attribute harus unik',
             'confirmed' => 'Konfirmasi password tidak cocok',
@@ -80,6 +87,17 @@ class MahasiswaController extends Controller
         ];
 
         $this->validate($request, $rules, $customMessages);
+
+        if ($request->foto) {
+            $value = $request->file('foto');
+            $extension = $value->extension();
+            $imageNames = uniqid('img_', microtime()) . '.' . $extension;
+            Storage::putFileAs('public/assets/foto-profil', $value, $imageNames);
+            $thumbnailpath = storage_path('app/public/assets/foto-profil/' . $imageNames);
+            $img = Image::make($thumbnailpath)->resize(600, 800)->save($thumbnailpath);
+        }else{
+            $imageNames = '';
+        }
 
         $user = User::create([
             'nama' => $request->nama,
@@ -106,6 +124,7 @@ class MahasiswaController extends Controller
             'tinggi_badan' => $request->tinggi_badan,
             'berat_badan' => $request->berat_badan,
             'angkatan' => $request->angkatan,
+            'foto' => $imageNames
         ]);
 
         return redirect()->route('data-mahasiswa.index')->with(['success' => 'Berhasil Menambah Data Mahasiswa ' . $request->nama]);
@@ -152,10 +171,10 @@ class MahasiswaController extends Controller
             'tempat_lahir' => 'required|string|max:20',
             'tanggal_lahir' => 'required|date',
             'jenis_kelamin' => 'required|in:L,P',
-            'alamat' => 'required|string|max:50',
+            'alamat' => 'required|string|max:255',
             'nama_ayah' => 'required|string|max:40',
             'nama_ibu' => 'required|string|max:40',
-            'alamat_orang_tua' => 'required|string|max:50',
+            'alamat_orang_tua' => 'required|string|max:255',
             'pekerjaan_ayah' => 'required|string|max:20',
             'pekerjaan_ibu' => 'required|string|max:20',
             'no_hp' => 'required|numeric',
@@ -174,15 +193,16 @@ class MahasiswaController extends Controller
         ];
 
         $rules4 = [
-            'npm' => 'required|string|max:255|unique:users'
+            'npm' => 'required|string|max:9|min:9|unique:users'
         ];
 
         $customMessages = [
             'required' => 'Field :attribute wajib diisi',
             'string' => 'Field :attribute harus berupa string',
             'numeric' => 'Field :attribute harus berupa angka',
-            'max' => 'Field :attribute maksimal :size',
-            'digits' => 'Field :attribute maksimal :value digit',
+            'max' => 'Field :attribute maksimal :max',
+            'min' => 'Field :attribute maksimal :min',
+            'digits' => 'Field :attribute harus :digits digit',
             'email' => 'Field :attribute harus berupa email',
             'unique' => 'Field :attribute harus unik',
             'confirmed' => 'Konfirmasi password tidak cocok',
@@ -196,6 +216,17 @@ class MahasiswaController extends Controller
         }
 
         $item = Mahasiswa::findOrFail($id);
+
+        if ($request->foto) {
+            $value = $request->file('foto');
+            $extension = $value->extension();
+            $imageNames = uniqid('img_', microtime()) . '.' . $extension;
+            Storage::putFileAs('public/assets/foto-profil', $value, $imageNames);
+            $thumbnailpath = storage_path('app/public/assets/foto-profil/' . $imageNames);
+            $img = Image::make($thumbnailpath)->resize(600, 800)->save($thumbnailpath);
+        }else{
+            $imageNames = $item->foto;
+        }
 
         if ($request->email != $item->users->email) {
             $this->validate($request, $rules3, $customMessages);
@@ -231,6 +262,7 @@ class MahasiswaController extends Controller
             'tinggi_badan' => $request->tinggi_badan,
             'berat_badan' => $request->berat_badan,
             'angkatan' => $request->angkatan,
+            'foto' => $imageNames,
         ]);
 
         return redirect()->route('data-mahasiswa.index')->with(['success' => 'Berhasil Mengubah Data Mahasiswa ' . $request->nama]);;
@@ -247,46 +279,62 @@ class MahasiswaController extends Controller
         $item = Mahasiswa::findOrFail($id);
         $idd = $item->user_id;
 
-        $item->forceDelete();
+        $check = BeritaKomentar::where('user_id', $idd)->first();
+        $check2 = DiskusiTanyaJawab::where('user_id', $idd)->first();
+        $check3 = LokerTanyaJawab::where('user_id', $idd)->first();
+        $check4 = Diskusi::where('user_id', $idd)->first();
 
-        $user = User::where('id', $idd)->first();
-        $nama = $user->nama;
+        if ($check || $check2 || $check3 || $check4) {
+            return redirect()->back()->with(['error' => 'Data Mahasiswa Tidak Bisa Dihapus']);
+        } else {
+            $item->forceDelete();
 
-        $user->forceDelete();
+            $user = User::where('id', $idd)->first();
+            $nama = $user->nama;
 
-        return redirect()->route('data-mahasiswa.index')->with(['success' => 'Berhasil Menghapus Data Mahasiswa ' . $nama]);
+            $user->forceDelete();
+
+            return redirect()->route('data-mahasiswa.index')->with(['success' => 'Berhasil Menghapus Data Mahasiswa ' . $nama]);
+        }
     }
 
     public function change_to_alumni($id)
     {
         $item = Mahasiswa::findOrFail($id);
 
-        Alumni::create([
-            'user_id' => $item->user_id,
-            'agama' => $item->agama,
-            'tempat_lahir' => $item->tempat_lahir,
-            'tanggal_lahir' => $item->tanggal_lahir,
-            'jenis_kelamin' => $item->jenis_kelamin,
-            'alamat' => $item->alamat,
-            'nama_ayah' => $item->nama_ayah,
-            'nama_ibu' => $item->nama_ibu,
-            'alamat_orang_tua' => $item->alamat_orang_tua,
-            'pekerjaan_ayah' => $item->pekerjaan_ayah,
-            'pekerjaan_ibu' => $item->pekerjaan_ibu,
-            'no_hp' => $item->no_hp,
-            'golongan_darah' => $item->golongan_darah,
-            'tinggi_badan' => $item->tinggi_badan,
-            'berat_badan' => $item->berat_badan,
-            'angkatan' => $item->angkatan,
-        ]);
+        if ($item->agama != NULL || $item->tempat_lahir != NULL || $item->tanggal_lahir != NULL || $item->jenis_kelamin != NULL || $item->alamat != NULL || $item->nama_ayah != NULL || $item->nama_ibu != NULL || $item->alamat_orang_tua != NULL || $item->pekerjaan_ayah != NULL || $item->pekerjaan_ibu != NULL || $item->no_hp != NULL || $item->golongan_darah != NULL || $item->tinggi_badan != NULL || $item->berat_badan != NULL || $item->angkatan != NULL) {
+            Alumni::create([
+                'user_id' => $item->user_id,
+                'agama' => $item->agama,
+                'tempat_lahir' => $item->tempat_lahir,
+                'tanggal_lahir' => $item->tanggal_lahir,
+                'jenis_kelamin' => $item->jenis_kelamin,
+                'alamat' => $item->alamat,
+                'nama_ayah' => $item->nama_ayah,
+                'nama_ibu' => $item->nama_ibu,
+                'alamat_orang_tua' => $item->alamat_orang_tua,
+                'pekerjaan_ayah' => $item->pekerjaan_ayah,
+                'pekerjaan_ibu' => $item->pekerjaan_ibu,
+                'no_hp' => $item->no_hp,
+                'golongan_darah' => $item->golongan_darah,
+                'tinggi_badan' => $item->tinggi_badan,
+                'berat_badan' => $item->berat_badan,
+                'angkatan' => $item->angkatan,
+            ]);
 
-        $user = User::where('id', $item->user_id)->first();
-        $user->role = 'ALUMNI';
-        $user->save();
+            $user = User::where('id', $item->user_id)->first();
+            $user->role = 'ALUMNI';
+            $user->save();
 
-        $item->delete();
+            $item->delete();
 
-        return redirect()->route('data-alumni.index')->with(['success' => 'Berhasil Memindah Data ke Data Alumni ' . $user->nama]);
+            return redirect()->route('data-alumni.index')->with(['success' => 'Berhasil Memindah Data ke Data Alumni ' . $user->nama]);
+        } else {
+            return redirect()->back()->with(['error-pindah' => 'Lengkapi Data Mahasiswa Terlebih Dahulu']);
+        }
+
+
+
     }
 
     public function import(Request $request)
